add basic form of authorization

+++ tsundoku/blueprints/api/entries.py
@@ -1,8 +1,9 @@
 import json
 import typing
 
-from quart import Response, request, views
+from quart import abort, Response, request, views
 from quart import current_app as app
+from quart_auth import current_user
 
 
 class EntriesAPI(views.MethodView):
@@ -17,6 +18,9 @@ class EntriesAPI(views.MethodView):
             A dict or a list of dict containing
             the requested entry information.
         """
+        if not await current_user.is_authenticated:
+            return abort(401, "You are not authorized to access this resource.")
+
         if entry_id is None:
             async with app.db_pool.acquire() as con:
                 entries = await con.fetch("""
@@ -24,7 +28,7 @@ class EntriesAPI(views.MethodView):
                     FROM show_entry WHERE show_id=$1;
                 """, show_id)
 
-            return json.dumps([dict(record) for record in entries])   
+            return json.dumps([dict(record) for record in entries])
         else:
             async with app.db_pool.acquire() as con:
                 entry = await con.fetchrow("""
@@ -62,6 +66,9 @@ class EntriesAPI(views.MethodView):
             Single key: `success`. Value is True if success,
             False otherwise.
         """
+        if not await current_user.is_authenticated:
+            return abort(401, "You are not authorized to access this resource.")
+
         required_arguments = {"episode", "magnet"}
         await request.get_data()
         arguments = await request.form
@@ -109,6 +116,9 @@ class EntriesAPI(views.MethodView):
             Single key: `success`. Value is True if success,
             False otherwise.
         """
+        if not await current_user.is_authenticated:
+            return abort(401, "You are not authorized to access this resource.")
+
         async with app.db_pool.acquire() as con:
             await con.execute("""
                 DELETE FROM show_entry WHERE id=$1;
